Add files via upload
Changed the algorithm of infectious circles that infects not infecterd circles. Instead of taking an infected circle on the map and iterating through all the not infected circles and calculation the distance between the 2 to know if he is close or not. The new algorithm creates a 2d array the size of the map with a 0 or 1 as a value. Each grid represent a pixel and if the pixel is 0, it is not contagious, if it is 1, it is contagious. We update this map by creating an area with the shape of a circle with the radius predetermined by the user and we place these value in the 2d array. This area is now full of 1 and it has the shape of a circle. Then we iterate through all the not infected circles and check if their position is on a 0 or a 1 in the new array. If he is, then we calculate the probability of getting infected and update the circles attributes.

I also added a new personnalize schedule that takes the time in miliseconds and play a turn with the timer decided in the constant TIMER
0.05 ms seems to be the sweet spot for 200 circles. Might need to get the Timer higher to get more circles in the canvas and not have a sloppy animation